# refactor: optimize bulk transaction for speed